+++ JSGame.cpp
@@ -875,7 +875,7 @@ JSAPI_FUNC(my_checkCollision)
 		if(!pUnit1 || !pUnit2)
 			return JS_TRUE;
 
-		*rval = INT_TO_JSVAL(D2COMMON_CheckUnitCollision(pUnit1, pUnit2, nBitMask));
+		*rval = INT_TO_JSVAL(D2COMMON_CheckUnitCollision(pUnit1, pUnit2, (WORD)nBitMask));
 		return JS_TRUE;
 	}
